requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
gunicorn
